# JSON serialization refactoring and optimization